Update description of GRAPHRAG_CACHE_BASE_DIR in env_vars.md (#1213)
* Update description of GRAPHRAG_CACHE_BASE_DIR in env_vars.md

Clarified that `GRAPHRAG_CACHE_BASE_DIR` refers to the base directory path for cache files rather than reporting outputs. This improves the accuracy of the documentation and helps users understand the correct usage of this environment variable.

* Update description of `GRAPHRAG_CACHE_BASE_DIR`

Simplified the description of `GRAPHRAG_CACHE_BASE_DIR` to make it clearer. Changed "base directory path" to "base path" for conciseness.

---------

Co-authored-by: Alonso Guevara <alonsog@microsoft.com>